requirements.txt
langchain
openai
langchain_community
python-dotenv
streamlit
langchain-community
chromadb
pypdf
tiktoken
docarray